+++ PSF2D/dec.h
@@ -4,12 +4,12 @@
 #include "mesh2d.h"
 #include "decmesh2d.h"
 
-Eigen::SparseMatrix<float> hodge2(DECMesh2D& mesh,float area,bool dual=false);
-Eigen::SparseMatrix<float> hodge1(DECMesh2D& mesh,float area,bool dual=false);
-Eigen::SparseMatrix<float> hodge0(DECMesh2D& mesh,float area,bool dual=false);
+Eigen::SparseMatrix<double> hodge2(DECMesh2D& mesh,double area,bool dual=false);
+Eigen::SparseMatrix<double> hodge1(DECMesh2D& mesh,double area,bool dual=false);
+Eigen::SparseMatrix<double> hodge0(DECMesh2D& mesh,double area,bool dual=false);
 
-Eigen::SparseMatrix<float> derivative0(DECMesh2D& mesh,bool dual=false);
-Eigen::SparseMatrix<float> derivative1(DECMesh2D& mesh,bool dual=false);
+Eigen::SparseMatrix<double> derivative0(DECMesh2D& mesh,bool dual=false);
+Eigen::SparseMatrix<double> derivative1(DECMesh2D& mesh,bool dual=false);
